Move descriptor to shared
