+++ executor/setup.py
@@ -7,6 +7,6 @@ setup(
     url="https://github.com/MXNetEdge/benchmark-ai",
     package_dir={"": "src"},
     packages=find_packages("src"),
-    package_data={"util": ["ec2_instance_info.csv"], "executor": "*.yaml", "transpiler.templates": "*.yaml"},
+    package_data={"executor": "*.yaml", "transpiler.templates": "*.yaml"},
     entry_points={"console_scripts": ["start=executor.__main__:main"]},
 )
