--- executor/smthng.yml
@@ -1,94 +1,0 @@
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: b-action-id-1npryjf8
-  labels:
-    app: benchmark-ai
-    action-id: ACTION-ID-1npryjf8
-    client-id: CLIENT_ID_7ywsfg5i
-    created-by: executor
-spec:
-  template:
-    metadata:
-      labels:
-        app: benchmark-ai
-        action-id: ACTION-ID-1npryjf8
-        client-id: CLIENT_ID_7ywsfg5i
-        created-by: executor
-      annotations:
-        iam.amazonaws.com/role: data-puller
-    spec:
-      serviceAccountName: metrics-pusher
-      affinity:
-        podAntiAffinity:
-          requiredDuringSchedulingIgnoredDuringExecution:
-          - labelSelector:
-              matchExpressions:
-              - key: app
-                operator: In
-                values:
-                - benchmark-ai
-            topologyKey: kubernetes.io/hostname
-      initContainers: []
-      containers:
-      - name: benchmark
-        image: alpine
-        resources:
-          limits:
-            nvidia.com/gpu: 0
-        command:
-        - sleep 
-        - 10s
-        securityContext:
-          privileged: false
-        env:
-        - name: BENCHMARK_AI
-          value: fifo
-        - name: BENCHMARK_AI_FIFO_FILEPATH
-          value: /tmp/benchmark-ai/fifo
-        - name: BENCHMARK_AI_FIFO_MAX_WAIT_TIME
-          value: '60'
-        volumeMounts:
-        - mountPath: /tmp/benchmark-ai
-          name: benchmark-ai
-        - name: dshm
-          mountPath: /dev/shm
-      - name: sidecar
-        image: benchmarkai/metrics-pusher:d9fc6b6
-        env:
-        - name: BENCHMARK_AI_FIFO_FILEPATH
-          value: /tmp/benchmark-ai/fifo
-        - name: BACKEND
-          value: elasticsearch
-        - name: BACKEND_ARG_JOB_ID
-          value: b-action-id-1npryjf8
-        - name: BACKEND_ARG_HOSTNAME
-          valueFrom:
-            configMapKeyRef:
-              name: outputs-infrastructure
-              key: es_endpoint
-        - name: BACKEND_ARG_PORT
-          value: '443'
-        - name: POD_NAME
-          valueFrom:
-            fieldRef:
-              fieldPath: metadata.name
-        - name: POD_NAMESPACE
-          valueFrom:
-            fieldRef:
-              fieldPath: metadata.namespace
-        volumeMounts:
-        - mountPath: /tmp/benchmark-ai
-          name: benchmark-ai
-      nodeSelector:
-        beta.kubernetes.io/instance-type: local
-        node.type: bai-worker
-        failure-domain.beta.kubernetes.io/zone: test-zone
-      restartPolicy: Never
-      volumes:
-      - name: benchmark-ai
-        emptyDir: {}
-      - name: dshm
-        emptyDir:
-          medium: Memory
-  backoffLimit: 4
