+++ kafka-utils/setup.py
@@ -14,5 +14,6 @@ setup(
     url="https://github.com/MXNetEdge/benchmark-ai",
     package_dir={"": "src"},
     packages=find_packages("src"),
+    package_data={"bai_kafka_utils.executors.util": ["ec2_instance_info.csv"]},
     include_package_data=True,
 )
