+++ kafka-utils/src/bai_kafka_utils/executors/util/ec2_instance_info.csv
