+++ kafka-utils/test-environment.yml
@@ -5,3 +5,5 @@ channels:
 dependencies:
   - pytest
   - pytest-mock
+  - pytest-datadir
+  - toml 0.9.*
