+++ kafka-utils/tests/bai_kafka_utils/executors/test_descriptor.py
@@ -1,7 +1,6 @@
 import pytest
-from bai_kafka_utils.executors.descriptor import DescriptorError, Descriptor
 
-import util.ec2_instance_info as ec2_instance_info
+from bai_kafka_utils.executors.descriptor import DescriptorError, Descriptor
 
 
 @pytest.mark.parametrize("filename", ["missing_keys_descriptor.toml", "missing_section_descriptor.toml"])
@@ -33,23 +32,6 @@ def test_invalid_args_type(descriptor_as_dict, descriptor_config):
         Descriptor(descriptor_as_dict, descriptor_config)
 
 
-def test_find_data_source(descriptor, base_data_sources):
-    source = descriptor.find_data_source(base_data_sources[0]["src"])
-    assert source["path"] == base_data_sources[0]["path"]
-
-
-def test_get_num_installed_gpus_valid_gpu():
-    assert ec2_instance_info.get_instance_gpus(instance_type="p3.8xlarge") == 4
-
-
-def test_get_num_installed_gpus_invalid_gpu():
-    with pytest.raises(Exception):
-        ec2_instance_info.get_instance_gpus(instance_type="p3.18000xlarge")
-
-
-def test_get_num_installed_gpus_valid_cpu():
-    assert ec2_instance_info.get_instance_gpus(instance_type="c5.18xlarge") == 0
-
-
-def test_get_num_installed_gpus_invalid_cpu():
-    assert ec2_instance_info.get_instance_gpus(instance_type="c5.18000xlarge") == 0
+def test_find_data_source(descriptor):
+    source = descriptor.find_data_source("foo1")
+    assert source["path"] == "bar1"
