+++ kafka-utils/tests/bai_kafka_utils/executors/test_descriptor/minimal_descriptor.toml
